Add `apitest` package for testing service handlers (#3)
This is a small augmentation to the API framework intended to provide a
package for testing service handlers. Service handlers are normal
functions and can be invoked as normal functions, but `apitest` provides
some extra niceties:

* Request structs are validated and an API error is emitted in case
  they're invalid.

* Response structs are validated and an error returned in case they're
  invalid.

We may add more things down the road as they're needed as well. For
example, `apitest` might inject a valid looking IP address into context
to simulate a more realistic request for purposes that need an IP like
rate limiting. (Nothing we're doing needs this yet, so I didn't bother.)

Sample usage:

    endpoint := &testEndpoint{}
    resp, err := apitest.InvokeHandler(ctx, endpoint.Execute, &testRequest{ReqField: "string"})
    require.NoError(t, err)